Add bool in/out support to pass ONNX backend tests.
